old-tests/submission/Infraware/Forms/contents/Forms/keygen_setcustomvalidity.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_setCustomValidity</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <keygen name='key' id='keygen_id'></keygen>
	</form>
	<script>

		var keygen = document.getElementById("keygen_id");

		try
		{
			keygen.setCustomValidity("custom error");
			test(function() {
				assert_true(true, "calling of setCustomValidity method is successed.");
			});
		}
		catch (e) {
			test(function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
